refactor: move Notifier protocol from engine to app target
The Notifier protocol was declared in the engine target but had
zero engine-side consumers — Engine signals via callbacks
(onProfileApplied, onUnknownLocation, onDevicesEvaluated), and only
AppDelegate calls notify(...). PR #54 moved both concrete
implementations (UserNotificationsNotifier, AppleScriptNotifier) to
the app target; this PR finishes the move by relocating the
protocol itself.

Net effect: the engine target's Adapters/ directory shrinks by one
orphan file, and Notifiers.swift in the app target is now self-
contained (no `import AVPainReliever` needed).

Engine target remains free of UI-framework imports
(`grep -rn "import AppKit\\|import UserNotifications\\|import SwiftUI\\|import Combine" Sources/AVPainReliever/`
returns empty). Behavior unchanged. swift test: 183 passed.
